GPU: Cleanup: Move ITS and TPC specific code out of GPUWorkFlowSpec.cxx to separate files
